refactor(LottoUtil): prettyPrint 이동 (-> LottoOutput)
