+++ src/main/java/lotto/view/LottoOutput.java
@@ -4,6 +4,9 @@ import lotto.constant.LottoPrintMessage;
 import lotto.domain.Lotto;
 import lotto.domain.LottoRound;
 import lotto.dto.IssuedLottoDTO;
+import lotto.dto.LottoWinStatisticDTO;
+
+import java.util.List;
 
 public class LottoOutput {
     public static void printIssuedLotto(IssuedLottoDTO issuedLottoDTO) {
@@ -15,5 +18,21 @@ public class LottoOutput {
         }
 
         System.out.println();
+    }
+
+    public static String prettyPrint(int value) {
+        String formattedPart = "";
+        if ( value > 1000 ) {
+            formattedPart = prettyPrint(value / 1000);
+            value %= 1000;
+            formattedPart += ",";
+        }
+
+        String formattingPart = Integer.toString(value);
+        if (!formattedPart.isEmpty()) {
+            if (formattingPart.length() < 2) formattingPart = "0" + formattingPart;
+            if (formattingPart.length() < 3) formattingPart = "0" + formattingPart;
+        }
+        return formattedPart + formattingPart;
     }
 }
